skills/jjannet/gold-price-mcp/requirements.txt
mcp-server
httpx
